[not verified] relocate utils folder
